# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alimentalos API",
        "description": "Official Alimentalos API Documentation",
        "contact": {
            "email": "iantorres@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v.0.6-beta"
    },
    "servers": [
        {
            "url": "https://api.alimentalos.cl",
            "description": "Backend production server"
        }
    ],
    "paths": {
...
        "/devices/{device}/geofences/{geofence}/detach": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Detach geofence of device.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachDeviceGeofence",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "Unique identifier of device",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "geofence",
                        "in": "path",
                        "description": "Unique identifier of geofence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device detached to geofence successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/users/{user}/geofences/{geofence}/detach": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Detach geofence of user.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachUserGeofence",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Unique identifier of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "geofence",
                        "in": "path",
                        "description": "Unique identifier of geofence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detached to geofence successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/groups/{group}/geofences/{geofence}/detach": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Detach geofence of group.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachGroupGeofence",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "geofence",
                        "in": "path",
                        "description": "Unique identifier of geofence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group detached to geofence successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pets/{pet}/geofences/{geofence}/detach": {
            "post": {
                "tags": [
                    "Pets"
                ],
                "summary": "Detach geofence of pet.",
                "description": "Returns empty array as JSON response.",
                "operationId": "detachPetGeofence",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "description": "Unique identifier of pet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "geofence",
                        "in": "path",
                        "description": "Unique identifier of geofence",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pet detached to geofence successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/users/{user}/geofences": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get geofences of user.",
                "description": "Returns the geofences of user paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getUserGeofences",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Unique identifier of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Geofences retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/devices/{device}/geofences": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get geofences of device.",
                "description": "Returns the geofences of device paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getDeviceGeofences",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "Unique identifier of device",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Geofences retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/groups/{group}/geofences": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get geofences of group.",
                "description": "Returns the geofences of group paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getUserGeofences",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Unique identifier of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Geofences retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/pets/{pet}/geofences": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Get geofences of pet.",
                "description": "Returns the geofences of pet paginated by a default quantity, payload includes pagination links and stats.",
                "operationId": "getPetGeofences",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "description": "Unique identifier of pet",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Geofences retrieved successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}